requirements.txt
opencv-contrib-python-headless
ultralytics
# roboflow
numpy
python-dotenv
pymongo
Jinja2
Quart
Werkzeug
supervision
hypercorn
watchdog 
python-socketio
PyYAML
ipython
paho-mqtt

